Looking at why certain behaviors are happening when doing random flashing, and potential fixes.
This first commit is encapsulating the 2 main ways which we want to handle flashing.

Ultimately, I think we want to split the need for the random generation of training targets and flashing targets into two different calls.

This main update has included finishing the Fisher Yates algorithm to actually be used. I am not sure why it wasn't swapped over previously, but am using it now, with the stipulation that the random arrays sequentially won't have the same neighboring values.