Fixed some bug with the "Trial Ended" being sent way too early,
In the old version of Bessy, there weren't distinct coroutines to run, but a monolithic giant code. When swapping to coroutines we ended up with behavior where the "Trial Ended" was sent early, because we weren't waiting for the coroutine of flashing to finish. This has introduced a new stop/start coroutine in BCI Controller behavior and has fixedc this issue.